회원가입, 로그인 구현

+++ backend/cowdog_be/src/main/java/com/xy/entity/MemberInfo.java
@@ -35,7 +35,7 @@ public class MemberInfo {
 	private String hobby;
 	private double latitude;
 	private double longitude;
-	private double distance;
+	private int distance;
 	private String address;
 
 }
